Move the OCR engine implementations into Text-Grab.Core.Windows
Continuing Phase 4. Moved the clean tier identified after mapping the
whole OCR/HDR area: TessLang, GlobalLang, WinAiOcrLinesWords (all
already dependency-free beyond ILanguage/IOcrLinesWords, which moved
earlier), and WinRtOcrLinesWords.

WinRtOcrLinesWords needed one fix: its constructor called
OcrUtilities.GetBoundingRect() - an app-side extension method that
computes the union of a line's word rects but returns a WPF
System.Windows.Rect, which the constructor then converted back into a
WinRT Windows.Foundation.Rect. Since every input (OcrWord.BoundingRect)
is already a WinRT Rect, that was a pointless WPF round-trip and the
actual blocker to moving the file. Replaced it with a local private
helper that computes the same union directly in Windows.Foundation.Rect.
OcrUtilities.GetBoundingRect itself is untouched - other app code may
still use it.

Also extracted HocrReader + TessOcrLine (the last ~53 lines of
TesseractHelper.cs) into Text-Grab.Core - pure string/regex hOCR
parsing, already cleanly separated from the rest of the file, going to
plain Core (not .Windows) since it's not Windows-specific. The rest of
TesseractHelper.cs stays in the app; nothing else currently calls it.

Added Microsoft.WindowsAppSDK.AI to Text-Grab.Core.Windows.csproj for
WinAiOcrLinesWords' RecognizedText/RecognizedLine/RecognizedWord types.

Deferred (each documented in the plan with its specific blocker):
OcrUtilities.cs (the actual "pivotal cut" - 1061 lines, mixed
headless/UI-adapter code plus a hidden WPF dependency in
LoadBitmapFromFile), the TesseractHelper class itself (settings-write
side effect in GetTesseractPath), LanguageService.cs and
WindowsAiUtilities.cs (both settings+WPF coupled, each dragging in
several more untouched types), BarcodeUtilities.cs (blocked on
OcrOutput.CleanOutput() reading settings directly), and the HDR/WGC
capture code (separate area, mostly clean already).
