+++ Text-Grab.Core.Windows/Models/GlobalLang.cs
