+++ Text-Grab.Core.Windows/Models/WinAiOcrLinesWords.cs
